Split activity surface derivation builders
