+++ Sources/QuillCodeApp/WorkspaceActivityStatusLabel.swift
@@ -1,0 +1,11 @@
+enum ActivityStatusLabel {
+    static let checked = "Checked"
+    static let done = "Done"
+    static let failed = "Failed"
+    static let logged = "Logged"
+    static let optional = "Optional"
+    static let pending = "Pending"
+    static let queued = "Queued"
+    static let review = "Review"
+    static let running = "Running"
+}
